requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.2.25
django-filter==23.5
djangorestframework==3.15.1
importlib-metadata==6.7.0
Markdown==3.4.4
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.7.1
zipp==3.15.0
dotenv==0.0.5
gunicorn==21.2.0